requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0
django-cors-headers==3.11.0
django-environ==0.9.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
tz==0.2.2
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.9
